1am we meet again
my sleep schedule suffers once again
- add todo list (aka wishlist aka camdens ideas)
- change pins for approximately the 27th time
- remove the rp2040 version checker (well really it just moved and it was pretty useless anyway I just saw it in the docs)
- headline feature (and it's like 5 points down)...api is here!!
- it took many many hours but I'm very happy with it!
- anyways
- fix flash writing
- more detailed flashmap (the used I had planned for it never made it in lol)
- start work on gps (I also just ordered one)
- fix position of includes and licenses in header files
- better debug info for interfaces
- imu addresses now use #define where possible
- led now has pulse functionality (that in itself took like 1-2 hours lol)
- wifly now internally handles its own status (idk why it was still in tcp)
- misc formatting changes
- switch internal units to km (nm were a bit too large)
- pidtune is back to using doubles (more accurate tuning possibly, pico has crazy perf anyways)
- fix bug in auto mode (first wpt alt would not have been loaded)
- add gps status to mode switcher
- allow external setting of normal mode setpoints (for api)
- fix some very very weird formatting in normal mode
- update tune mode to utilize flight (that sounds dumb now that I read it again)
- fix some possible bugs in tune mode

phew, that REALLY was a lot!!

+++ src/modes/hold.c
@@ -21,7 +21,7 @@ static uint8_t turnStatus = HOLD_TURN_UNSCHEDULED;
 
 static float oldHeading;
 static float targetHeading;
-static int_fast16_t targetAlt;
+static int targetAlt;
 
 static double rollSet;
